installing vim before script

+++ install.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
+#install vim & vim-runtime
+sudo apt install vim vim-runtime -y
 
 #vim directory
 mkdir ~/.vim && mkdir ~/.vim/autoload
-
+#download vim plug
 git clone https://github.com/tpope/vim-pathogen.git > ~/.vim/autoload/
 
 #install plugin manager
@@ -12,8 +14,11 @@ curl -fLo ~/.vim/autoload/plug.vim --create-dirs\
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 
 #copy vimrc to ~/.vimrc and save old config to ~/.vimrc.save
-mv ~/.vimrc ~/.vimrc.save
-mv vimrc ~/.vimrc
+if [ -e ~/.vimrc ] then
+	mv ~/.vimrc ~/.vimrc.save
+fi
+
+mv ./vimrc ~/.vimrc
 
 #install plugins
 vim -c 'PlugInstall'
